+++ sklearn/cluster/affinity_propagation_.py
@@ -11,6 +11,7 @@ import numpy as np
 
 from ..base import BaseEstimator, ClusterMixin
 from ..utils import as_float_array, check_array
+from ..utils.validation import check_is_fitted
 from ..metrics import euclidean_distances
 from ..metrics import pairwise_distances_argmin
 
@@ -313,9 +314,7 @@ class AffinityPropagation(BaseEstimator, ClusterMixin):
         labels : array, shape (n_samples,)
             Index of the cluster each sample belongs to.
         """
-        if not hasattr(self, "cluster_centers_indices_"):
-            raise ValueError("Estimator is not fitted.")
-
+        check_is_fitted(self, "cluster_centers_indices_")
         if not hasattr(self, "cluster_centers_"):
             raise ValueError("Predict method is not supported when "
                              "affinity='precomputed'.")