+++ sklearn/lda.py
@@ -21,6 +21,7 @@ from .linear_model.base import LinearClassifierMixin
 from .covariance import ledoit_wolf, empirical_covariance, shrunk_covariance
 from .utils.multiclass import unique_labels
 from .utils import check_array, check_X_y
+from .utils.validation import check_is_fitted
 from .preprocessing import StandardScaler
 
 
@@ -441,6 +442,8 @@ class LDA(BaseEstimator, LinearClassifierMixin, TransformerMixin):
         X_new : array, shape (n_samples, n_components)
             Transformed data.
         """
+        check_is_fitted(self, ['xbar_', 'scalings_'], all_or_any=any)
+
         X = check_array(X)
         if self.solver == 'lsqr':
             raise NotImplementedError("transform not implemented for 'lsqr' "